remove nav from movies provider
