+++ client/src/components/Layout/index.js
@@ -1,19 +1,35 @@
 import React from 'react'
+import { useRouteMatch } from 'react-router-dom'
+
 import { useAuth } from '../../auth'
 import NavigationHeader from '../NavigationHeader'
 
+function getPreviousUrlInHistory({ isHouseDetailRoute }) {
+  if (isHouseDetailRoute) {
+    return '/houses'
+  }
+}
+
 const Layout = ({ children }) => {
   const [logged] = useAuth()
+  const isHouseDetailRoute = useRouteMatch('/houses/:houseName')?.isExact
 
-  if (!logged) {
+  const previousUrlInHistory = getPreviousUrlInHistory({ isHouseDetailRoute })
+
+  if (logged) {
     return (
       <>
-        <NavigationHeader background="#D70808" color={'white'} />
+        <NavigationHeader backLink={previousUrlInHistory} />
         {children}
       </>
     )
   }
-  return children
+  return (
+    <>
+      <NavigationHeader background="#D70808" color={'white'} />
+      {children}
+    </>
+  )
 }
 
 export default Layout
